optimize memory footprint and simplify filterbank initialization
- Compact cb_width arrays in SR_INFO to uint8_t and sbr_offset to int8_t.
- Update tns_sfb_range struct bounds to use fixed-width uint8_t types.
- Remove inner const qualifiers from hcode16_t and hcode32_t definitions.
- Flatten duplicate loop logic in FillKbdWindow for Kaiser-Bessel windows.
- Inline sine/KBD window initialization in FilterBankInit to drop WindowPair.

+++ libfaac/coder.h
@@ -16,6 +16,7 @@
 #ifndef CODER_H
 #define CODER_H
 
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -105,8 +106,8 @@ typedef struct {
   unsigned long sampling_rate;  /* the following entries are for this sampling rate */
   int num_cb_long;
   int num_cb_short;
-  int cb_width_long[NSFB_LONG];
-  int cb_width_short[NSFB_SHORT];
+  uint8_t cb_width_long[NSFB_LONG];
+  uint8_t cb_width_short[NSFB_SHORT];
 } SR_INFO;
 
 /* Scalefactor-band layout per sampling_rate_index, shared by frame.c and sbr.c. */
